requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
opencv-python==4.7.0.72
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.2.2
tensorflow==2.13.0
torch==2.0.0
torchvision==0.15.0
Pillow==9.5.0
matplotlib==3.7.1
seaborn==0.12.2
reportlab==4.0.4
python-magic==0.4.27
pillow-heif==0.14.0
librosa==0.10.0
soundfile==0.12.1
werkzeug==2.3.0
requests==2.31.0
python-dotenv==1.0.0
